restructure test cases for full strategy and enable time index.
