+++ hc-dna/zomes/tests/full-time-index/subject-object-link.ts
@@ -1,4 +1,4 @@
-import { localConductorConfig, installation } from './common'
+import { localConductorConfig, installation } from '../common'
 
 module.exports = (orchestrator) => {
 	orchestrator.registerScenario("Subject object link test", async (s, t) => {
